requirements.txt
notebook==7.3.1
pandas==2.2.3
pillow==11.0.0
pillow-avif-plugin==1.4.6
PyPDF2==3.0.1
transformers==4.47.1
torch==2.5.1